Issue 82335: Moved parameter into its own function, added doc fragment
